big refactoring

+++ src/text_selection_app/statistics.py
@@ -3,29 +3,28 @@ from logging import getLogger
 from pathlib import Path
 from typing import cast
 
-from ordered_set import OrderedSet
 import pandas as pd
-from text_selection_core.statistics import generate_statistics
+from ordered_set import OrderedSet
 
-from text_selection_app.argparse_helper import (ConvertToOrderedSetAction,
-                                                parse_existing_directory,
-                                                parse_non_empty_or_whitespace)
+from text_selection_app.argparse_helper import (ConvertToOrderedSetAction, parse_existing_directory,
+                                                parse_non_empty, parse_non_empty_or_whitespace)
 from text_selection_app.helper import get_datasets
-from text_selection_app.io_handling import (get_data_n_grams_path,
-                                            get_data_symbols_path,
-                                            get_data_weights_path, load_data_n_grams,
-                                            load_data_symbols, load_data_weights,
-                                            load_dataset)
+from text_selection_app.io_handling import get_data_weights_path, load_data_weights, load_dataset
+from text_selection_core.statistics import generate_statistics
 
 
 def get_statistics_generation_parser(parser: ArgumentParser):
-  parser.description = f"This command creates statistics."
+  parser.description = "This command creates statistics."
   parser.add_argument("directory", type=parse_existing_directory, metavar="directory",
                       help="directory containing data")
   parser.add_argument("--weights", type=parse_non_empty_or_whitespace, nargs="*", metavar="NAME",
                       help="name of the weights", default=[], action=ConvertToOrderedSetAction)
-  parser.add_argument("--n-grams", type=parse_non_empty_or_whitespace, nargs="*", metavar="NAME",
-                      help="name of the weights", default=[], action=ConvertToOrderedSetAction)
+  parser.add_argument("file", type=parse_non_empty_or_whitespace,
+                      help="name of the file")
+  parser.add_argument("--lsep", type=parse_non_empty, default="\n",
+                      help="line separator")
+  parser.add_argument("--ssep", type=str, default="",
+                      help="symbol separator")
   parser.add_argument("-o", "--overwrite", action="store_true",
                       help="overwrite")
   return statistics_generation_ns
@@ -43,14 +42,6 @@ def statistics_generation_ns(ns: Namespace) -> None:
                     ) if root_folder != data_folder else "root"
     logger.info(f"Processing {data_name} ({i}/{len(datasets)})")
 
-    symbols_path = get_data_symbols_path(data_folder)
-    if not symbols_path.exists():
-      logger.error(
-        f"Symbols were not found! Ignoring...")
-      symbols = None
-    else:
-      symbols = load_data_symbols(symbols_path)
-
     weights = []
     for weights_name in cast(OrderedSet[str], ns.weights):
       weights_path = get_data_weights_path(data_folder, weights_name)
@@ -61,19 +52,16 @@ def statistics_generation_ns(ns: Namespace) -> None:
       current_weights = load_data_weights(weights_path)
       weights.append((weights_name, current_weights))
 
-    n_grams = []
-    for n_grams_name in cast(OrderedSet[str], ns.n_grams):
-      n_grams_path = get_data_n_grams_path(data_folder, n_grams_name)
-      if not n_grams_path.is_file():
-        logger.error(
-          f"N-Grams '{n_grams_name}' were not found! Ignoring...")
-        continue
-      current_n_grams = load_data_n_grams(n_grams_path)
-      n_grams.append((n_grams_name, current_n_grams))
+    symbols_path = data_folder / cast(str, ns.file)
+    if not symbols_path.exists():
+      logger.error(
+        "File was not found! Skipping...")
+      continue
+    lines = symbols_path.read_text(ns.encoding).split(ns.lsep)
 
     dataset = load_dataset(dataset_path)
     logger.debug("Generating statistics...")
-    dfs = generate_statistics(dataset, symbols, weights, n_grams)
+    dfs = generate_statistics(dataset, lines, ns.ssep, weights)
     stats_path = root_folder / "statistics.csv"
     header_indicator = "###"
     with open(stats_path, mode="w") as f:
